refactor: Move all types and interfaces to the types.ts folder and add imports where needed
