+++ app/lib/data/answerKey.ts
@@ -12,7 +12,6 @@ export const correctKeySigNotationAnswers = [
 
 export const correctKeySigAnswers = ["db", "a", "f", "c#"];
 
-//need to update the logic to allow for either 7 or 8 notes
 const scaleDbMajor = ["db", "eb", "f", "gb", "ab", "bb", "c"];
 const scaleBMajor = ["b", "c#", "d#", "e", "f#", "g#", "a#"];
 const scaleCDorian = ["c", "d", "eb", "f", "g", "a", "bb"];
