+++ app/lib/isClickWithinStaveBounds.ts
@@ -14,14 +14,6 @@ const isClickWithinStaveBounds = (
   const minLeftClick = measureWidth * 0.08;
   const minTopClick = topStaveMaxYClick;
   const maxBottomClick = bottomStaveMaxYClick;
-  // console.log(
-  //   `measure width: ${measureWidth}\n`,
-  //   `max right click: ${maxRightClick}\n`,
-  //   `min left click: ${minLeftClick}\n`,
-  //   `min top click: ${minTopClick}\n`,
-  //   `max bottom click: ${maxBottomClick}`
-  // );
-
   const isWithinBounds = !(
     userClickX < minLeftClick ||
     userClickX > maxRightClick ||
